// types.ts
export interface SyncMetadata {
  _deleted?: boolean;
  updatedAt?: string; // ISO String
}

export interface BankData extends SyncMetadata {
  id: string;
  name: string;
  promoRate: number;
  fixedYear: number;
  type: 'KPR' | 'KKB' | 'KTA';
}

// NEW: Bank Account Entity
export interface BankAccount extends SyncMetadata {
  id: string;
  userId: string;
  bankName: string;
  accountNumber: string;
  holderName: string;
  balance: number;
  color: string; // Hex or tailwind class
  type: 'Bank' | 'E-Wallet' | 'Cash';
}
// ...
export interface User {
  id: string;
  username: string;
  email: string;
  password?: string; 
  role: 'admin' | 'user';
  status: 'active' | 'pending_verification' | 'inactive';
  lastLogin?: string; 
  parentUserId?: string | null; 
  createdAt: string;
  updatedAt?: string; // NEW: Track profile updates
  photoUrl?: string; 
  sessionToken?: string; 
  
  riskProfile?: 'Conservative' | 'Moderate' | 'Aggressive';
  bigWhyUrl?: string; 
  financialFreedomTarget?: number; 
  badges?: string[]; 
  
  // Analytics Fields (Admin Dashboard)
  totalDebt?: number;
  totalIncome?: number;
  monthlyObligation?: number;
}

export interface Badge {
  id: string;
  name: string;
  description: string;
  icon: string;
  color: string;
}